add cache in playlist

+++ src/services/postgres/PlaylistsService.js
@@ -6,8 +6,9 @@ const AuthorizationError = require("../../exceptions/AuthorizationError");
 const { mapDBToModel, mapDBToModelPlaylists } = require("../../utils");
 
 class PlaylistsService {
-  constructor() {
+  constructor(cacheService) {
     this._pool = new Pool();
+    this._cacheService = cacheService;
   }
 
   async addPlaylist({ name, owner }) {
@@ -24,25 +25,37 @@ class PlaylistsService {
       throw new InvariantError("Playlist gagal ditambahkan");
     }
 
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username
+    try {
+      // mendapatkan playlists dari cache
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username
       FROM playlists
       LEFT JOIN users ON users.id = playlists.owner
       LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
       WHERE playlists.owner = $1 OR collaborations.user_id = $1
       GROUP BY playlists.id, users.username
-      
-      
     `,
-      values: [owner],
-    };
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModelPlaylists);
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModelPlaylists);
+
+      await this._cacheService.set(
+        `playlists:${owner}`,
+        JSON.stringify(mappedResult)
+      );
+
+      return mappedResult;
+    }
   }
 
   async getSongById(id) {
@@ -59,23 +72,9 @@ class PlaylistsService {
     return result.rows.map(mapDBToModel)[0];
   }
 
-  async editSongById(id, { title, year, performer, genre, duration }) {
-    const updatedAt = new Date().toISOString();
-    const query = {
-      text: `UPDATE songs SET title = $1, year = $2, performer = $3, genre = $4, duration = $5, "updatedAt" = $6 WHERE id = $7 RETURNING id`,
-      values: [title, year, performer, genre, duration, updatedAt, id],
-    };
-
-    const result = await this._pool.query(query);
-
-    if (!result.rowCount) {
-      throw new NotFoundError("Gagal memperbarui Lagu. Id tidak ditemukan");
-    }
-  }
-
   async deletePlaylistById(id) {
     const query = {
-      text: "DELETE FROM playlists WHERE id = $1 RETURNING id",
+      text: "DELETE FROM playlists WHERE id = $1 RETURNING id, owner",
       values: [id],
     };
 
@@ -84,6 +83,9 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError("Lagu gagal dihapus. Id tidak ditemukan");
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 
   async verifyPlaylistOwner(id, owner) {
